% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2_individual.R
\name{ptTWAS_stage2_individual}
\alias{ptTWAS_stage2_individual}
\title{Stage 2: Regression with GWAS Outcome}
\usage{
ptTWAS_stage2_individual(
  Z,
  G_til,
  model_stage1,
  outcome = "binary",
  simul_band_cutoffs = c(0.05, 0.04, 0.03, 0.02, 0.01, 5e-05, 0.001),
  boot_iter = 1000
)
}
\arguments{
\item{Z}{A numeric vector of GWAS outcome.}

\item{G_til}{A numeric matrix of genotypes, with individuals in rows and SNPs
in columns. Row names are the subject IDs.}

\item{model_stage1}{The gene expression model fitted by \code{ptTWAS_stage1}.}

\item{outcome}{Either `binary` or `continuous`.}

\item{simul_band_cutoffs}{A numeric vector of alpha levels to test for the simultaneous confidence band.}

\item{boot_iter}{Number of bootstrap iterations.}
}
\value{
A list containing the results of the Stage 2 regression:
   \item{`alpha_hat`}{The estimate of vector \eqn{\alpha} under orthonormal basis splines.}
   \item{`alpha_var`}{The variance-covariance estimate of vector \eqn{\alpha} under orthonormal basis splines.}
   \item{`Global_pval`}{P-value for global test.}
   \item{`nonflat_pval`}{P-value for nonflat test.}
   \item{`gamma_til`}{The calibrated confidence factor.}
}
\description{
This function performs the second stage regression (scalar-on-function) in pseudotime-dependent
Transcriptome-Wide Association Study (pt-TWAS) using individual-level GWAS data. Still, it uses continuous orthonormal splines to model the
gene effect curve \eqn{\alpha}(t).
}
\details{
This function regress the GWAS outcome onto predicted gene expression curve (scalar-on-function regression).
Inferences are based on hypothesis tests (global, non-flat) and simultaneous confidence band.
}
