% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2_summary.R
\name{ptTWAS_stage2_summary}
\alias{ptTWAS_stage2_summary}
\title{Stage 2: Regression with GWAS Outcome}
\usage{
ptTWAS_stage2_summary(
  GWAS_beta_hat,
  GWAS_beta_se,
  GWAS_sample_size,
  outcome = "binary",
  model_stage1,
  ref_panel,
  control_case_ratio = NULL,
  af_vec = NULL,
  var_Z = NULL
)
}
\arguments{
\item{GWAS_beta_hat}{GWAS regression coefficients.}

\item{GWAS_beta_se}{GWAS regression coefficient sds.}

\item{GWAS_sample_size}{GWAS sample size.}

\item{outcome}{Either `binary` or `continuous`.}

\item{model_stage1}{The gene expression model fitted by \code{ptTWAS_stage1}.}

\item{ref_panel}{A reference panel.}

\item{control_case_ratio}{Control-case ratio. Only available when \code{outcome='binary'}.}

\item{af_vec}{Allele frequencies of SNPs. If \code{NULL}, estimated from the reference panel.}

\item{var_Z}{Variance of outcome \code{Z}. Only available when \code{outcome='continuous'}.}
}
\value{
A list containing the results of the Stage 2 regression:
   \item{`alpha_hat`}{The estimate of vector \eqn{\alpha} under orthonormal basis splines.}
   \item{`alpha_var`}{The variance-covariance estimate of vector \eqn{\alpha} under orthonormal basis splines.}
   \item{`Global_pval`}{P-value for global test.}
   \item{`nonflat_pval`}{P-value for nonflat test.}
}
\description{
This function performs the second stage regression (scalar-on-function) in pseudotime-dependent
Transcriptome-Wide Association Study (pt-TWAS) using summary-level GWAS data. Still, it uses continuous orthonormal splines to model the
gene effect curve \eqn{\alpha}(t).
}
\details{
This function regress the GWAS outcome onto predicted gene expression curve (scalar-on-function regression) using summary statistics.
Inferences are based on hypothesis tests (global, non-flat) only.
}
